+++ src/mdx/_TerraMdxProvider.jsx
@@ -3,6 +3,7 @@ import { MDXProvider } from '@mdx-js/react';
 import PropTypes from 'prop-types';
 import Tag from './_Tag';
 import HTag from './_HTag';
+import Code from './_Code';
 
 const propTypes = {
   /**
@@ -18,7 +19,7 @@ const propTypes = {
 const components = {
   a: props => (Tag({ Tag: 'a', props })),
   blockquote: props => (Tag({ Tag: 'blockquote', props })),
-  code: props => (Tag({ Tag: 'code', props })),
+  // code: props => (Tag({ Tag: 'code', props })),
   dd: props => (Tag({ Tag: 'dd', props })),
   dl: props => (Tag({ Tag: 'dl', props })),
   dt: props => (Tag({ Tag: 'dt', props })),
@@ -42,6 +43,7 @@ const components = {
   th: props => (Tag({ Tag: 'th', props })),
   tr: props => (Tag({ Tag: 'tr', props })),
   ul: props => (Tag({ Tag: 'ul', props })),
+  code: props => (Code({ ...props })),
 };
 
 /**